fix local variable resolved index in block
unlike original Lox, we see block as another expression,
thus it's possible for us to declare new variable (locals),
in the process of doing binary expressions.

in original Lox, such operations are statements, and they
leave nothing on the stack, thus it doesn't have any issue.

the solution is to wrap binary expression, and count the
stack height, and add that when resolving local variables